+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/akkastreams/dispatcher/DispatcherImpl.scala
@@ -1,6 +1,6 @@
 // Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
-//TODO: move the tests as well
+
 package com.digitalasset.platform.akkastreams.dispatcher
 
 import java.util.concurrent.atomic.AtomicReference
