+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/platform/akkastreams/dispatcher/DispatcherTest.scala
@@ -1,14 +1,13 @@
 // Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.akkastreams
+package com.digitalasset.platform.akkastreams.dispatcher
 
 import java.util.concurrent.atomic.{AtomicInteger, AtomicReference}
 
 import akka.NotUsed
 import akka.stream.scaladsl.{Keep, Sink}
 import com.digitalasset.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.digitalasset.platform.akkastreams.dispatcher.Dispatcher
 import com.digitalasset.platform.akkastreams.dispatcher.SubSource.OneAfterAnother
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.time.{Milliseconds, Seconds, Span}
@@ -16,7 +15,8 @@ import org.scalatest.{Matchers, WordSpec}
 
 import scala.concurrent.{ExecutionContextExecutor, Future}
 
-class DispatcherIT extends WordSpec with AkkaBeforeAndAfterAll with Matchers with ScalaFutures {
+//TODO: merge/review the tests we have around the Dispatcher!
+class DispatcherTest extends WordSpec with AkkaBeforeAndAfterAll with Matchers with ScalaFutures {
 
   override implicit def patienceConfig: PatienceConfig =
     PatienceConfig(scaled(Span(10, Seconds)), scaled(Span(250, Milliseconds)))
